Add eslint plugin to detect implicit dependencies
Implicit dependencies are where code calls `require('modulename')` but without having `modulename` listed as a dependency of any kind. This generally occurs when said module is already installed as a child of something else in the dependency tree, so is present in ./node_modules.

However, this causes code to become fragile as it is dependent on the exact structure of the dependency tree, which is liable to change uncontrollably and unexpectedly. In particular, if a module is loaded as a dev dependency (or descendant of a dev dependency) it will not be present following a production install and so may cause code that has passed CI to fail in a prod environment.

By ensuring that all dependencies are explicitly declared in package.json we can remove this risk.